知识库关联表改造成通过id

+++ server/db/repository/knowledge_metadata_repository.py
@@ -1,3 +1,6 @@
+from sqlalchemy import func
+
+from server.db.models.knowledge_base_model import KnowledgeBaseModel
 from server.db.models.knowledge_metadata_model import SummaryChunkModel
 from server.db.session import with_session
 from typing import List, Dict
@@ -5,14 +8,15 @@ from typing import List, Dict
 
 @with_session
 def list_summary_from_db(session,
-                         kb_name: str,
+                         kb_id: int,
                          metadata: Dict = {},
                          ) -> List[Dict]:
     '''
     列出某知识库chunk summary。
     返回形式：[{"id": str, "summary_context": str, "doc_ids": str}, ...]
     '''
-    docs = session.query(SummaryChunkModel).filter(SummaryChunkModel.kb_name.ilike(kb_name))
+
+    docs = session.query(SummaryChunkModel).filter(SummaryChunkModel.kb_id == kb_id)
 
     for k, v in metadata.items():
         docs = docs.filter(SummaryChunkModel.meta_data[k].as_string() == str(v))
@@ -21,7 +25,7 @@ def list_summary_from_db(session,
              "summary_context": x.summary_context,
              "summary_id": x.summary_id,
              "doc_ids": x.doc_ids,
-             "metadata": x.metadata} for x in docs.all()]
+             "metadata": x.meta_data} for x in docs.all()]
 
 
 @with_session
@@ -32,11 +36,13 @@ def delete_summary_from_db(session,
     删除知识库chunk summary，并返回被删除的Dchunk summary。
     返回形式：[{"id": str, "summary_context": str, "doc_ids": str}, ...]
     '''
-    docs = list_summary_from_db(kb_name=kb_name)
-    query = session.query(SummaryChunkModel).filter(SummaryChunkModel.kb_name.ilike(kb_name))
+    kb: KnowledgeBaseModel = session.query(KnowledgeBaseModel).filter(KnowledgeBaseModel.kb_name == kb_name).first()
+    if kb is None:
+        return list()
+    query = session.query(SummaryChunkModel).filter(SummaryChunkModel.kb_id == kb.id)
     query.delete(synchronize_session=False)
     session.commit()
-    return docs
+    return list()
 
 
 @with_session
@@ -47,9 +53,12 @@ def add_summary_to_db(session,
     将总结信息添加到数据库。
     summary_infos形式：[{"summary_context": str, "doc_ids": str}, ...]
     '''
+    kb: KnowledgeBaseModel = session.query(KnowledgeBaseModel).filter(KnowledgeBaseModel.kb_name == kb_name).first()
+    if kb is None:
+        return True
     for summary in summary_infos:
         obj = SummaryChunkModel(
-            kb_name=kb_name,
+            kb_id=kb.id,
             summary_context=summary["summary_context"],
             summary_id=summary["summary_id"],
             doc_ids=summary["doc_ids"],
@@ -63,4 +72,7 @@ def add_summary_to_db(session,
 
 @with_session
 def count_summary_from_db(session, kb_name: str) -> int:
-    return session.query(SummaryChunkModel).filter(SummaryChunkModel.kb_name.ilike(kb_name)).count()
+    kb: KnowledgeBaseModel = session.query(KnowledgeBaseModel).filter(KnowledgeBaseModel.kb_name == kb_name).first()
+    if kb is None:
+        return 0
+    return session.query(func.count(SummaryChunkModel.id)).filter(SummaryChunkModel.kb_id == kb.id).scalar()
